requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0.1
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-js-asset==1.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
